Merge commit 'a630642585cc' from llvm.org/main into next (#12012)
Conflicts:
      clang/lib/Analysis/UnsafeBufferUsage.cpp

rdar://166521282
(cherry picked from commit d060dacf2435fb11871cc030176414a9cf07524c)

Message of the merged commit 'a630642585cc':
[-Wunsafe-buffer-usage] Check isValueDependent before EvaluateAsBooleanCondition (#172091)

The function `EvaluateAsBooleanCondition` assumes (asserts) that the
input `Expr` is not in a dependent context. So it is caller's
responsibility to check the condition before the call. This commit fixes
the issue in `UnsafeBufferUsage.cpp`.

The issue was first found downstream because of some code not
upstreamed. This commit also includes those un-upstreamed code.

rdar://166217941

Co-authored-by: Ahmed Bougacha <ahmed@bougacha.org>