[lldb] Prefer exact address match when looking up symbol by address (#172055)
The current behavior will pick the first symbol that contains the
address, this causes LLDB to pick the wrong symbol when looking for
swift reflection metadata on Linux, as in that case it is valid for a
symbol to completely encompass another one.

Instead, this function should prefer the symbol which is an exact, if it
exists. As a bonus, this should also be faster in the vast majority of
the cases, as we probably query symbols by their exact address most of
the time.

rdar://166344740
(cherry picked from commit 208553460a704fc94b5154c0a2145010015b137f)